+++ src/Manager/ShipTerminal/TerminalCommand.cs
@@ -3,5 +3,5 @@
 namespace QualityCompany.Manager.ShipTerminal;
 
 [AttributeUsage(AttributeTargets.Method)]
-internal class TerminalCommand : Attribute
+public class TerminalCommand : Attribute
 { }